refactor(nuget): move functionalitiy in dedicated files
